+++ src/test/resources/Http_GET.txt
@@ -1,0 +1,4 @@
+GET /user/create?userId=javajigi&password=password&name=JaeSung HTTP/1.1
+Host: localhost:8080
+Connection: keep-alive
+Accept: */*
